requirements.txt
# Django and REST Framework
Django==5.2.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9  # For PostgreSQL (production)

# Filtering and CORS
django-filter==24.2
django-cors-headers==4.3.1

# API Documentation
coreapi==2.3.3
Pygments==2.17.2
drf-yasg==1.21.7  # Better API documentation with Swagger

# Development and Testing
pytest==8.2.2
pytest-django==4.8.0
pytest-cov==5.0.0  # Test coverage for pytest
factory-boy==3.3.0
coverage==7.5.3
django-test-plus==2.2.2  # Enhanced testing utilities

# Environment variables
python-decouple==3.8
django-environ==0.11.2  # Alternative environment management

# Caching (Redis)
redis==5.0.6
django-redis==5.4.0

# Celery for background tasks
celery==5.3.6
kombu==5.3.7  # Message transport for Celery

# Database utilities
django-extensions==3.2.3  # Useful Django extensions
django-debug-toolbar==4.4.2  # Debug toolbar for development

# Data validation and processing
jsonschema==4.22.0  # JSON schema validation
python-dateutil==2.9.0  # Enhanced date parsing

# Excel/CSV processing
openpyxl==3.1.2  # Excel file processing
xlsxwriter==3.2.0  # Excel file writing
pandas==2.2.2  # Data analysis and CSV processing

# API rate limiting
django-ratelimit==4.1.0

# Permissions and security
django-guardian==2.4.0  # Object-level permissions (future use)
cryptography==42.0.8  # Cryptographic functions

# AI/ML Libraries (for future chatbot development)
tensorflow==2.16.1
scikit-learn==1.5.0
nltk==3.8.1
numpy==1.26.4

# Text-to-Speech (from existing PDF code)
pyttsx3==2.90

# Data visualization
matplotlib==3.9.0
seaborn==0.13.2
plotly==5.22.0  # Interactive plots

# Image processing
Pillow==10.3.0

# HTTP requests (for external API integration)
requests==2.32.3
httpx==0.27.0  # Async HTTP client

# Task scheduling (alternative to Celery)
django-crontab==0.7.1

# Email utilities
django-ses==4.0.0  # AWS SES email backend

# Development tools
black==24.4.2
flake8==7.0.0
isort==5.13.2
pre-commit==3.7.1  # Git hooks for code quality
mypy==1.10.0  # Type checking
django-stubs==5.0.2  # Django type stubs for mypy

# Production server and deployment
gunicorn==22.0.0
whitenoise==6.7.0
django-storages==1.14.4  # File storage backends (S3, etc.)
boto3==1.34.131  # AWS SDK

# Monitoring and logging
sentry-sdk==2.5.0
django-silk==5.1.0  # Performance profiling
structlog==24.2.0  # Structured logging

# API client tools (for testing)
httpie==3.2.2

# Documentation generation
sphinx==7.3.7  # Documentation generator
sphinx-rtd-theme==2.0.0  # Read the Docs theme

# Backup and fixtures
django-fixture-generator==1.0.0  # Generate fixtures from data

# Health checks
django-health-check==3.18.2

# Internationalization
django-modeltranslation==0.18.11  # Model translation (future use)

# Background job processing alternatives
django-q==1.6.1  # Alternative to Celery
rq==1.16.2  # Redis Queue

# Memory profiling (development)
memory-profiler==0.61.0

# Social authentication (future use)
django-allauth==0.63.3

# API versioning
django-rest-framework-version-transforms==0.3.0

# Additional utilities
python-slugify==8.0.4  # URL-friendly slugs
qrcode==7.4.2  # QR code generation
reportlab==4.2.0  # PDF generation

# Timezone handling
pytz==2024.1